Gracefully handle empty db in clean method

+++ src/test-helpers/index.js
@@ -27,7 +27,9 @@ export const getDBHelper = (connectionString: string) => {
       // For some reason flow thinks that `result` might be Promise<ResultSet> | ResultSet
       // $FlowFixMe
       const result = await getDatabase().query(sql)
-
+      if (!result.rows.length) {
+        return
+      }
       return await getDatabase().query(
         `TRUNCATE TABLE ${result.rows
           .map(table => `${schema}.${table['table_name']}`)
